pyproject.toml
[tool.ruff]
line-length = 120
[project]
name = "backendcourse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "black>=25.1.0",
    "fastapi-cache2>=0.2.2",
    "fastapi[standard]>=0.115.13",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-dotenv>=0.5.2",
    "redis>=6.2.0",
    "ruff>=0.12.9",
    "sqlalchemy>=2.0.41",
]


